Provide a warp feature.
All my warp + ructe projects use the same RenderRucte extension trait
to make calling the templates on generating responses a bit clearer.
Provide that trait here as an optional feature.
