+++ Cargo.toml
@@ -13,7 +13,7 @@ license = "MIT/Apache-2.0"
 [features]
 sass = ["rsass"]
 mime02 = []
-mime03 = []
+mime03 = ["mime"]
 
 [dependencies]
 base64 = "^0.10.0"
@@ -24,6 +24,8 @@ md5 = "0.6"
 nom = { version = "4.1.1", features = ["verbose-errors"] }
 
 rsass = { version = "0.9.8", optional = true }
+warp = { version = "0.1.9", optional = true }
+mime = { version = "0.3", optional = true }
 
 [badges]
 travis-ci = { repository = "kaj/ructe" }
