+++ examples/warp/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 build = "src/build.rs"
 
 [build-dependencies]
-ructe = { path = "../..", features = ["sass", "mime03"] }
+ructe = { path = "../..", features = ["warp", "sass", "mime03"] }
 
 [dependencies]
 warp = "0.1.9"
